refactor: :recycle: refactor: organizando as pastas e importações e o readME.
