+++ src/routes/authRoute.js
@@ -1,7 +1,7 @@
 import { Router } from "express";
 import AuthController from "../controllers/authController.js";
 import { celebrate } from 'celebrate';
-import { gerenciadorDeErros, validacaoAuth } from '../middlewares/index.js'
+import { gerenciadorDeErros, validacaoAuth } from '../utils/index.js'
 
 const router = Router();
