+++ src/utils/validacao/cargosValidacao.js
@@ -1,5 +1,5 @@
 import { Joi, Segments } from 'celebrate';
-import { mensagensCustomizadas } from '../../utils/validacoesMensagens/mensagensCustomizadas.js';
+import { mensagensCustomizadas } from '../validacoesMensagens/mensagensCustomizadas.js';
 
 const validacaoCargos = {
     [Segments.BODY]: Joi.object().keys({
